victoriametrics: optimize API for quering tags
Since 1.47.0, VictoriaMetrics implements /tags/autoComplete API
which is way more faster than using same approach as was performed
for prometheus.
Use it when it's available on a backend.

Related to #521
